requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
asyncpg
httpx
requests
passlib[bcrypt]  
bcrypt==4.0.0
python-jose>=3.0.0
python-multipart